refactor(usage): group the breakdown's sub-components
Three of them, so they move into components/ and drop the prefix the path
already supplies: Row, List, ScopeTotal.

The pair also matches UsageDashboard now. The bare name is the container, so
the folder, the file and the default export agree, and callers stop reaching
past the barrel for it.
