+++ frontend/web/components/organisation-settings/usage/UsageBreakdown/UsageBreakdown.tsx
@@ -1,56 +1,91 @@
-import { FC } from 'react'
-import UsageBreakdownList from './UsageBreakdownList'
-import { BREAKDOWN_DIMENSIONS, BreakdownDimension, BreakdownRow } from './utils'
-import './UsageBreakdown.scss'
+import { FC, useMemo, useState } from 'react'
+import { skipToken } from '@reduxjs/toolkit/query'
+import { Req } from 'common/types/requests'
+import { Res } from 'common/types/responses'
+import { useGetProjectsQuery } from 'common/services/useProject'
+import { useGetEnvironmentsQuery } from 'common/services/useEnvironment'
+import UsageBreakdownView from './UsageBreakdownView'
+import ScopeTotal, { UsageScope } from './components/ScopeTotal'
+import { useScopedBreakdown } from './useScopedBreakdown'
+import { byRequestType, bySdk, BreakdownDimension } from './utils'
 
 export type UsageBreakdownProps = {
-  dimension: BreakdownDimension
-  onChangeDimension: (dimension: BreakdownDimension) => void
-  rows: BreakdownRow[]
-  isLoading?: boolean
-  /** Environments belong to a project, so the dimension needs one chosen. */
-  needsProject?: boolean
+  organisationId: number
+  billingPeriod: Req['getOrganisationUsage']['billing_period']
+  /** Already fetched for the charts above, so these two cost nothing extra. */
+  data: Res['organisationUsage'] | undefined
+  /** The page's project filter. Environments can only be listed under one. */
+  projectId: number | undefined
 }
 
-type DimensionOption = (typeof BREAKDOWN_DIMENSIONS)[number]
-
-/**
- * Where the usage came from. Deliberately carries no plan limit: a single
- * request type or project has no allowance of its own, so a limit line here
- * would invite a comparison that means nothing.
- *
- * Totals only, no series over time. The chart above answers "where am I
- * heading against my plan", and a second time chart here would compete with it.
- */
 const UsageBreakdown: FC<UsageBreakdownProps> = ({
-  dimension,
-  isLoading,
-  needsProject,
-  onChangeDimension,
-  rows,
-}) => (
-  <div className='p-4 mt-3 border border-default rounded-lg bg-surface-default'>
-    <div className='d-flex align-items-center justify-content-between gap-3 mb-3'>
-      <strong>Where the usage came from</strong>
-      <div className='usage-breakdown__dimension'>
-        <Select
-          onChange={(option: DimensionOption) =>
-            onChangeDimension(option.value)
-          }
-          value={BREAKDOWN_DIMENSIONS.find(
-            (option) => option.value === dimension,
-          )}
-          options={BREAKDOWN_DIMENSIONS}
-        />
-      </div>
-    </div>
+  billingPeriod,
+  data,
+  organisationId,
+  projectId,
+}) => {
+  const [dimension, setDimension] = useState<BreakdownDimension>('request-type')
 
-    <UsageBreakdownList
-      rows={rows}
-      isLoading={isLoading}
-      needsProject={needsProject}
-    />
-  </div>
-)
+  const { data: projects } = useGetProjectsQuery(
+    dimension === 'project' ? { organisationId } : skipToken,
+  )
+  const { data: environments } = useGetEnvironmentsQuery(
+    dimension === 'environment' && projectId ? { projectId } : skipToken,
+  )
+
+  const scopes: UsageScope[] = useMemo(() => {
+    if (dimension === 'project') {
+      return (projects ?? []).map((project) => ({
+        key: `project-${project.id}`,
+        label: project.name,
+        projectId: project.id,
+      }))
+    }
+    if (dimension === 'environment') {
+      return (environments?.results ?? []).map((environment) => ({
+        environmentId: environment.api_key,
+        key: `environment-${environment.api_key}`,
+        label: environment.name,
+      }))
+    }
+    return []
+  }, [dimension, projects, environments])
+
+  const scoped = useScopedBreakdown(
+    scopes,
+    `${dimension}|${billingPeriod ?? 'rolling'}|${projectId ?? 'all'}`,
+  )
+
+  const rows = useMemo(() => {
+    if (dimension === 'request-type') return byRequestType(data)
+    if (dimension === 'sdk') return bySdk(data)
+    return scoped.rows
+  }, [dimension, data, scoped.rows])
+
+  const needsProject = dimension === 'environment' && !projectId
+
+  return (
+    <>
+      {!needsProject &&
+        scopes.map((scope) => (
+          <ScopeTotal
+            key={scope.key}
+            organisationId={organisationId}
+            billingPeriod={billingPeriod}
+            scope={scope}
+            onTotal={scoped.onTotal}
+          />
+        ))}
+
+      <UsageBreakdownView
+        dimension={dimension}
+        onChangeDimension={setDimension}
+        rows={rows}
+        isLoading={scoped.isLoading}
+        needsProject={needsProject}
+      />
+    </>
+  )
+}
 
 export default UsageBreakdown
